@@ -4,15 +4,15 @@ apply plugin: 'maven-publish'
 group = 'com.github.haiwen'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 33
 
     defaultConfig {
         minSdkVersion 16
-        targetSdkVersion 27
+        targetSdkVersion 31
         versionCode 1
-        versionName "0.19.0"
+        versionName "0.19.2"
 
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
 
     }
     buildTypes {
@@ -26,10 +26,12 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
 
-    implementation 'com.android.support:appcompat-v7:27.0.1'
+    final def x_version = '1.5.0'
+    implementation "androidx.appcompat:appcompat:$x_version"
+
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.1'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.1'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.0'
     //Flexmark
     implementation 'com.vladsch.flexmark:flexmark:0.28.32'
     implementation 'com.vladsch.flexmark:flexmark-ext-tables:0.28.32'
@@ -41,7 +43,7 @@ dependencies {
     implementation 'com.vladsch.flexmark:flexmark-ext-footnotes:0.28.32'
     implementation 'com.vladsch.flexmark:flexmark-ext-attributes:0.28.32'
     //Logger
-    implementation 'com.orhanobut:logger:1.15'
+//    implementation 'com.orhanobut:logger:1.15'
 }
 
 afterEvaluate {
@@ -51,7 +53,7 @@ afterEvaluate {
                 from components.release
                 groupId = 'com.github.haiwen'
                 artifactId = 'MarkdownView'
-                version = '0.19.1'
+                version = '0.19.2'
             }
         }
     }
