@@ -1,6 +1,7 @@
 package br.tiagohm.markdownview.ext.bean.internal;
 
-import com.orhanobut.logger.Logger;
+import android.util.Log;
+
 import com.vladsch.flexmark.html.CustomNodeRenderer;
 import com.vladsch.flexmark.html.HtmlWriter;
 import com.vladsch.flexmark.html.renderer.NodeRenderer;
@@ -51,11 +52,12 @@ public class BeanNodeRenderer implements NodeRenderer {
                         fieldMethod = beanClass.getMethod("is" + methodNameCamelCase);
                         //Logger.d("encontrado is%", methodNameCamelCase);
                     } catch (NoSuchMethodException e2) {
-                        Logger.d("NoSuchMethodException: ", methodName);
+                        Log.d(BeanNodeRenderer.class.getName(), "NoSuchMethodException: " + methodName);
                         return null;
                     }
                 }
             }
+
             if (fieldMethod != null) {
                 Object o;
                 try {
